Insert routes and stops into database.
